Remove outdated files
Initially the intention was that we'd use this repository as a
cookiecutter style template, however that never materialized and the
need was reduced by adopting the client monorepo.

Since then we've adopted ruff in place of black and isort, so those
files are now outdated. The dependabot one is still up to date, but
simple enough that we don't need to maintain it here.